+++ mirror/create_mirror_misc.sh
@@ -34,7 +34,8 @@ curl -LOJ https://artifacts.elastic.co/downloads/logstash/logstash-5.2.2.tar.gz
 tar zxf logstash-5.2.2.tar.gz
 rm logstash-5.2.2.tar.gz
 cd logstash-5.2.2
-bin/logstash-plugin install $PLUGIN_LIST
+# work around bug introduced in 5.1.1: https://discuss.elastic.co/t/5-1-1-plugin-installation-behind-proxy/70454
+JARS_SKIP='true' bin/logstash-plugin install $PLUGIN_LIST
 bin/logstash-plugin prepare-offline-pack $PLUGIN_LIST
 chmod a+r logstash-offline-plugins-5.2.2.zip
 mv logstash-offline-plugins-5.2.2.zip $STATIC_FILE_DIR/logstash-offline-plugins-5.2.2.zip